% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphSensitivity.R
\name{add}
\alias{add}
\title{Visualizing sensivity results}
\usage{
add(x, y)
}
\arguments{
\item{x}{Count of pairs where only control has outcome.}

\item{y}{Count of pairs where onlytreated has outcome.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Visualizing sensivity results
}
\examples{
add (1, 1)
add (10, 1)
}

