% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphSensivity.R
\name{graphSensitivity}
\alias{graphSensitivity}
\title{Visualizing sensivity results}
\usage{
graphSensitivity(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Visualizing sensivity results
}
\examples{
subtract(1, 1)
subtract (10, 1)
}

