% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphSensivity.R
\name{subtract}
\alias{subtract}
\title{Subtract two numbers.}
\usage{
subtract(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Subtract two numbers.
}
\examples{
subtract(1, 1)
subtract (10, 1)
}

